Move control styles next to source
