+++ TF_Settings_Web/src/Components/Controls/TabSelector/TabSelector.scss
