+++ TF_Settings_Web/src/Components/Controls/ToggleSwitch/ToggleSwitch.scss
