+++ TF_Settings_Web/src/Components/Controls/ToggleSwitch/ToggleSwitch.tsx
@@ -1,4 +1,4 @@
-import 'Styles/Controls/ToggleSwitch.scss';
+import './ToggleSwitch.scss';
 
 import React from 'react';
