+++ TF_Settings_Web/src/Components/Pages/Camera/ManualSetup/ManualSetupScreen.tsx
@@ -1,3 +1,5 @@
+import { TextEntry } from '@/Components/Controls/TextEntry/TextEntry';
+
 import './ManualSetup.scss';
 
 import { Component } from 'react';
@@ -6,8 +8,6 @@ import { ConfigurationManager } from 'TouchFree/src/Configuration/ConfigurationM
 import { PhysicalConfig } from 'TouchFree/src/Configuration/ConfigurationTypes';
 import { ConnectionManager } from 'TouchFree/src/Connection/ConnectionManager';
 import { ConfigState, WebSocketResponse } from 'TouchFree/src/Connection/TouchFreeServiceTypes';
-
-import { TextEntry } from 'Components/Controls/TextEntry';
 
 interface PhysicalState {
     screenHeight: number;
