+++ TF_Settings_Web/src/Components/Pages/Camera/QuickSetup/Calibration/CalibrationCompleteScreen.tsx
@@ -1,3 +1,5 @@
+import { TextButton } from '@/Components/Controls/TFButton/TFButton';
+
 import './Calibration.scss';
 import cssVariables from 'Styles/variables.module.scss';
 
@@ -6,8 +8,6 @@ import { useNavigate } from 'react-router-dom';
 
 import { InputActionManager } from 'TouchFree/src/Plugins/InputActionManager';
 import { InputType, TouchFreeInputAction } from 'TouchFree/src/TouchFreeToolingTypes';
-
-import { TextButton } from 'Components/Controls/TFButton';
 
 import { CalibrationHandLostMessage, CalibrationPracticeButton } from './CalibrationComponents';
